WebSocket Gateway reorganization
- remove chat/game seperate gateways/modules
- use a generic "socket" module

* See README.md in "backend/src/socket" for more details.
