--- backend/src/chat/chat.controller.ts
@@ -1,4 +1,0 @@
-import { Controller } from '@nestjs/common';
-
-@Controller('chat')
-export class ChatController {}
