--- backend/src/chat/chat.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { ChatGateway } from './chat.gateway';
-import { ChatController } from './chat.controller';
-
-@Module({
-  controllers: [ChatController],
-  providers: [ChatGateway],
-})
-export class ChatModule {}
