--- backend/src/game/game.gateway.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { GameGateway } from './game.gateway';
-
-describe('GameGateway', () => {
-  let gateway: GameGateway;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [GameGateway],
-    }).compile();
-
-    gateway = module.get<GameGateway>(GameGateway);
-  });
-
-  it('should be defined', () => {
-    expect(gateway).toBeDefined();
-  });
-});
