+++ backend/src/socket/socket.module.ts
@@ -1,0 +1,7 @@
+import { Module } from '@nestjs/common';
+import { SocketGateway } from './socket.gateway';
+
+@Module({
+  providers: [SocketGateway]
+})
+export class SocketModule {}
